requirements.txt
fastapi>=0.115.5
uvicorn[standard]>=0.32.1
langgraph>=0.2.50
langchain-google-genai>=2.0.4
langchain-core>=0.3.15
python-dotenv>=1.0.1
pydantic>=2.9.2
httpx>=0.27.2
google-generativeai>=0.8.3